Font size stuff should actually, really, honestly work properly now.
Changed various functions in DasherViewSquare.inl to use screen coordinates
rather that dasher ones as it's not an absolute mapping any more. This
probably wants to be optimised.

The screen is created before fontsize is set, by the looks of things. That
whole section needs to be dealt with properly.