Apply Gentoo dasher-4.6.1-gnome.patch permit compilation without gnome.
The patch uncommented several "#ifdef GNOME_LIBS" lines that were
noted as being potentially defective in r2987.  Since there are a
number of other references to GNOME_LIBS in other parts of the source,
I have to assume the issue has been resolved.  If not, the situation
now is no worse than it was before this patch, and the #ifdefs are now
more consistent.


svn path=/trunk/; revision=3489