Encourage bitrate change to have near immediate effect.
When you increase the framerate, the new value of iSteps (which has to
be much smaller, i.e. perform the expansion to the cursor over iSteps
frames) is interpolated with the previous value, so it takes several
updates to get to where it needs to be. This patch recomputes iSteps
on receipt of the speed change event(s).