g/c actions
Remove the action code (+ extra commented out code) #ifdef'd out
in 4c37ab02. It looks as though the code was the basis for a dasher
scripting language. Sadly it also seems as though the complete
source only existed on a past developer's computer and was never
committed to the repository in its entirety.