Fix event dispatch loop
The loop was based on iterating through a deque of event handlers.
The problem is that an event handler, such as is called when changing
the alphabet, can cause new observer objects to be created, which
register their handlers by adding them to the deque. Unfortunately
according to the C++ standard

  23.3.3.4 deque modifiers

  Effects: An insertion in the middle of the deque invalidates all
  the iterators and references to elements of the deque. An insertion
  at either end of the deque invalidates all the iterators to the
  deque, but has no effect on the validity of references to elements
  of the deque.

so the iterators become invalid during the event dispatch loop.

g++ is extremely lenient, allowing push_back additions to the deque,
but other compilers aren't - afterall I could implement push_back
by malloc(oldsize + 1) and copy, so any pointers to elements in
the old array would not point to current new array locations and
be invalid.