Feature/124_linux ci setup (#140)
Travisci errors out:
> config.status: error: cannot find input file: `Makefile.in'

It seems that the generated Makefile refers to the README file for some reason. So we need to add the original README file back to the repository for now. We need to figure out where exeatly README is referenced and see how we can update it to README.md or delete the reference.


* [124] Remove README file

README file is replaced by README.md file.

* Removed the duplication in README.md

* Add README back